+++ packages/iov-ethereum/types/wsethereumrpcclient.d.ts
@@ -1,0 +1,12 @@
+import { Stream } from "xstream";
+import { JsonRpcRequest, JsonRpcResponse } from "@iov/jsonrpc";
+import { SocketWrapperMessageEvent } from "@iov/socket";
+import { EthereumRpcClient } from "./ethereumrpcclient";
+export declare class WsEthereumRpcClient implements EthereumRpcClient {
+    readonly events: Stream<SocketWrapperMessageEvent>;
+    private readonly socket;
+    constructor(baseUrl: string);
+    run(request: JsonRpcRequest): Promise<JsonRpcResponse>;
+    socketSend(request: JsonRpcRequest, ignoreNetworkError?: boolean): Promise<void>;
+    disconnect(): void;
+}